prism_source/prism/requirements.txt
Django==4.2.6
asgiref==3.7.2
django-rest-framework==0.1.0
djangorestframework== 3.14.0
pip==21.1.2
pytz==2023.3.post1
setuptools==57.0.0
sqlparse==0.4.4
tzdata==2023.3
wheel==0.36.2
mysqlclient==2.1.1
djangorestframework-simplejwt==5.3.0